pyproject.toml
[tool.poetry]
name = "afc-squad-bot"
version = "0.1.0"
description = ""
authors = ["Afc_Adinho <afcadinho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
